I added ALL the ItemType enumerations.
Yes, I added them all. They're all there now, including all the potions and
all the new 1.8 items. Everything is there.

On a more minor note, I also had to change the getData() method in ItemType
(plus the interface HoldableType and BlockType) to return a short instead
of a byte because potion data values are too big.

Also, I removed that list of changes between the site and the enumerations
from the Javadoc in BlockType and didn't put any in ItemType. I actually
used a different site for some of the ItemTypes because the site I usually
use removed all the potions, plus I didn't think anyone would really care
that much and it just cluttered the Javadoc.

MAN, we got a lot done today! I love weekends!

+++ Corundum/utils/interfaces/HoldableType.java
@@ -18,7 +18,7 @@ public interface HoldableType<T> extends IDedType<T> {
      * 
      * @return the data value associated with this {@link IDedType} or -1 if this {@link IDedType} is not associated with a particular data value like {@link IDedType#GRASS
      *         grass blocks}. */
-    public byte getData();
+    public short getData();
 
     /** This method returns the maximum number of this type of item that can be put into one single inventory slot. Many items and all blocks have a max stack size of 64.
      *
